Create a single disk which has fast get/search.
GET on average requires reading one memory page from disk + the contents
of the query.

SEARCH requires reading < 12MB to figure out where else to read to
satisfy queries.  Search must be done in an iterative fashion (e.g., you
cannot read all from disk at once and cache in memory or try to blast
over the network).  Creating an iterator to the disk is a pointer copy,
integer copy, and atomic increment.  Iterating disk requires reading a
portion of the 12MB, and then has the cost of a GET.