Flush the data to disk.
Commit 4c729f8a63f59d56e611f20eced0b635f1c3701b implemented sub-regions,
but did not flush to disk (data was left in a multi-producer,
single-consumer iterable list).  Commit
e9bd89481563faa0723968228e698f116b85ed59 called a flush method on
individual regions, but this method was a NOP.  This commit is the code
that complements these two to truly flush to disk.  It had a slight bug
in the use of reference-counted pointers which needed to be resolved
first.