Properly(?) step the head iterator.
When flushing or destructing the log, the head iterator was stepped by
self assignment.  The e::intrusive_ptr should be able to do this on its
own, but the rest of the code uses a temporary.  By using a temporary
here, we play it safe in case it is indeed necessary.