Found the bug that caused the occasional NOTFOUND.
The 'count' client issues alternating PUT/GET requests for keys.  By the
guarantees of linearizability, we know that the GET should ALWAYS return
the PUT.  This was not the case.  It happened ~1/1000 times with the
default disk-flushing timeouts, and more often with more aggressive
timings.

The problem:  For certain keys, if the key was put to disk, it would be
dropped because the network-exposed regions created disks with the same
prefix/mask.  This is incorrect because network-exposed regions'
coordinates are created solely by the dimensions which form the region's
subspace, while the on-disk regions are formed using all dimensions.
Thus, it was incorrect to use the same region for both.  The solution is
to use a region of (X, Y, 0, 0) instead, as it will capture all elements
from the network region.