Add a PENDING message type.
A PENDING message is sent along the point leader's chain in reverse.  It
signals that a PUT or DEL is now pending on every node that it must
touch and is waiting for an ACK from the point leader.  The receipt of
this message type indicates that a host must (TODO) mark the message as
such so that if the host becomes point-leader, it can immediately
commit/ACK the message so that the client does not see a prefix of the
true history.