Add special functions to the messaging layer.
The special (e.g., not entity to entity) functions take extra parameters
to determine the entity to which (or from which) a message is sent.  For
example, if we know that we must send an ACK from a region, we can just
send(region, ...) and let the communication layer sort out the proper
entity which fits as the source (or it drops the message).  The three
special functions are:

 - forward in chain A, or to the head of B
 - forward in chain A, or to the tail of B
 - backward in chain A, or to the tail of B

These replace the ad-hoc variations of these, and make it easier to
understand exactly where a message will go (the ability to send to head
or tail of B when sending forward in a chain was prompted by the
upcoming change referred to in the previous commit RE: PENDING).