Yay!  We have linearizability!
Up until this point our replication scheme didn't provide
linearizability in the case of point-leader failures.

The specific case we had trouble with was:
1.  The last replica in subspace n-1 sends an update message to the
    point leader.
2.  The point leader sends ACK to client.
3.  The point leader dies and the client message is not lost.  The
    second replica in subspace 0 becomes a point leader, and has no clue
    that the message was acknowledged.
4.  The client sends a get to the new point leader and sees a break from
    our linearizability guarantees.

The fix:  After sending to all replicas in all subspaces, extra PENDING
message[1] is sent in reverse along the entire subspace 0 chain,
starting at the last replica and working toward the point-leader.  The
point leader may commit/ack the client when it receives this message,
and may then send the ACK message as it did before.

1.  Commit 12358c8c8d1b4b82ed44deb7adb90ebfcbb65d03