Stop-the-world reconfiguration.
There is a critical section in which we can safely assert that network
threads will not be processing messages.  The critical section also
guarantees that any messages processed after it will use the new
configuration for every step of the processing for the message (this is
just one state transition for the whole update -- it is entirely
possible that a message is marked pending under the old configuration,
but acknowledged under the new configuration.