Snapshots work.
The snapshotted entity can iterate all valid items on a disk which were
present at the time the snapshot was taken, and will not be affected by
put or del operations on the disk which occur after the snapshot.  The
cost of taking a snapshot?  Copying two pointers and a 32-bit integer
within a critical section.