Take snapshots of a region.
A "snapshot" takes a snapshot at an instantaneous point in time.  No
amount of changes to the region will cause the snapshot to return
different results than would be returned if no changes were made to the
region.

A "rolling_snapshot" takes a "snapshot" and also grabs an iterator to
the region's in-memory log.  It will return everything the snapshot
does, and then every item which exists in the log (in order).