Use binary search to update search index.
There are 2**18 entries, which means we now do at most 18 checks rather
than a significantly larger number.  Where this helps is in the flusher
thread.  The one thread which flushes to disk does the invalidation, and
during heavy read/write workloads, latency went out the window because
the writes would back up in the in-memory WAL, and so each GET would
need to traverse the entire log.