Do a fair round-robin when polling.
The array of file descriptors was set up in increasing order, and
checked beginning to end after a poll.  This unfairly biased
communication to favor those connections with a low descriptor.  This
patch makes it so that communication is relatively fair and balanced by
picking up where a previous poll left off.