Refactor replication code.
The code did not clearly distinguish between messages which only touch one host
per subspace and messages which need to touch two in a subspace.  Indeed, if
there were more than one subspace in which the update needed to touch two hosts,
the thing would break.  I've clearly marked where the code to handle this case
should go, and will include it next as I wanted to make the changes to the
working code first.