Send CHAIN_SUBSPACE messages to all replicas.
A CHAIN_SUBSPACE message indicates that the point is moving from one set
of replicas in a subspace to another.  Before CHAIN_PUT was being sent
to all the other replicas instead of CHAIN_SUBSPACE.  This was incorrect
as it did not allow the replica to properly place the update (it was
marked deferred instead of pending).