Change the hash lookup for shards.
The old method would overwrite dead entries in the hash table with
different keys.  The new method preserves "one bucket == one key".

Included in this patch is a fix for a bug which caused shard::copy_to to
produce bad results.  It would leave the hash table on the destination
empty while inserting duplicate entries into the hash table on the
source.