Fix a bug where deletes were dropped.
When putting data to shards, we need to remove from the old shard, and
add to the new shard.  As an optimization, we only performed the DEL on
the old shard if it was not the same as the new shard.  Unfortunately, I
forgot to check that we had indeed done a PUT when taking the
optimization and so we would lose DELs.

This was caught by tests the moment I committed the bad code, I just
didn't run the tests to catch it.