Remove buffer usage in the hyperspacehashing lib.
Buffers are heavyweight and memory-owning.  In order to pass the key or
value to an object, a new buffer had to be created, and the contents
copied.  This buffer could be passed by reference into the hashing
layer, but the damage of a memory allocation/copy was already done.

This patch uses the new e "slice" object which refers to memory in a
lightweight, non-memory-owning manner.  The rest of the app (built on
hsh) can now follow this pattern.

The reason for this change was simple:  For a single key and value,
there were, on average, 14 buffer copies per operation, all of which are
unnecessary.  As this change percolates up the code base, it will
eliminate all of these useless operations.  The result will be reduced
CPU usage (right now, a number of clients can saturate one host's CPU).