Add some invariant checks for CLIENT_PUT case.
When a client does a CLIENT_PUT, and that operation does not specify
every attribute, it needs to preserve the unspecified attributes.  It
does so by allocating more memory and copying them.  We neglected to
copy the key, so it was pointing into unused memory.  I fixed this bug,
and added assertions to make sure it won't happen again.