Fix the bug in search code.
Breakdown:  In the rush to fix performance of the client, I moved to
indexing outstanding requests using the (fd, nonce).  The old method
didn't depend upon the nonce.  When a search asks for the next item, the
nonce changes.  The fix is to update the index of outstanding requests
to reflect the new nonce.