Fix a bug affecting spaces with no search attrs.
If a space did not have any attributes marked for search on the local disk, the
mask would be 0.  When searching, it would then match anything, regardless of
the mask.  When sectors split in four, the set with the bit set would be matched
when the others should have been matched.