Include key size in calculations of microop sizes.
An OP_INT64_ADD may be for a map or a normal number.  In the former
case, we'd neglect the key size, and corrupt memory.

We fix this by pessimistically assuming the worst for every map op.
This will be faster, but allocate slightly more memory than is strictly
necessary.