Automate the transfer process from beginning to end
When "initiate-transfer" is called, the tail of the region sends all state to
the new node.  Once all state is sent, the coordinator is asked to make the
transfer "live" by appending the node to the end of the region's chain.
Finally, once all state is transferred, the tail node is made a full member of
the chain and the transfer is complete.