Fix a deadlock between background and network threads.
We paused the background threads by holding a lock across reconfiguration.
Unfortunately, we did so by first blocking the threads and then the network
traffic.  If a packet arrived that required passing info to the background
threads (by holding their lock), we could deadlock.  Fix this by adding another
set of condition variables between the background threads and the thread that
performs reconfiguration.