Use an anonymous BusyBee ID for the HyperDex client
The ID generation was not robust enough to be safe when launching hundreds of
clients simultaneously, occasionally leading to clients with duplicate IDs.
This would lead one of the two clients to hang, as its replies would get sent to
the other by any daemons to which they both connected.